migrations: fix data types
This commit fixes data types for the foreign keys and adds tutorial
on how to work with migrations (creating and deleting
tables structure of the database) and working with mysql database.

+++ database/migrations/2022_11_01_225019_create_exercises_table.php
@@ -15,7 +15,7 @@ class CreateExercisesTable extends Migration
     {
         Schema::create('exercises', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('author');
+            $table->unsignedBigInteger('author');
             $table->string('name');
             $table->string('description');
             $table->string('topic');
